refactor create-user logic
