+++ backend/src/libs/services/user.services.ts
@@ -1,10 +1,36 @@
+import bcrypt from 'bcryptjs';
 import { HttpError } from '@libs/api-gateway';
 import UserModel from '@libs/model/user.model';
+import { ICreateUserReqBody } from '@libs/types/createUserReqBody.type';
+import { IUser } from '@libs/types/user.type';
 
 class UserServices {
   private userModel: UserModel;
   constructor() {
     this.userModel = new UserModel();
+  }
+
+  public async createUser(userId: string, reqBody: ICreateUserReqBody) {
+    const { firstName, lastName, email, password, isAdmin, message, allergy } =
+      reqBody;
+
+    const encryptedPW = bcrypt.hashSync(password.trim(), 10);
+
+    const userData: IUser = {
+      PK: 'user',
+      SK: userId,
+      firstName,
+      lastName,
+      email,
+      password: encryptedPW,
+      isAdmin,
+      message,
+      allergy,
+    };
+
+    console.log('createUser userData', userData);
+
+    await this.userModel.createUser(userData);
   }
 
   public async errorIfUserNotExist(
@@ -28,6 +54,14 @@ class UserServices {
       eventId,
       isAttending
     );
+  }
+
+  public async errorIfUserExists(email: string) {
+    const existingUser = await this.userModel.getUserByEmail(email);
+
+    if (existingUser.Items.length > 0) {
+      throw new HttpError(500, 'User already exists');
+    }
   }
 }
