+++ backend/src/libs/types/createUserReqBody.type.ts
@@ -1,0 +1,9 @@
+export interface ICreateUserReqBody {
+  firstName: string;
+  lastName: string;
+  email: string;
+  password: string;
+  isAdmin: boolean;
+  message: string;
+  allergy: string;
+}
