+++ backend/src/libs/types/user.type.ts
@@ -1,0 +1,11 @@
+export interface IUser {
+  PK: string;
+  SK: string;
+  firstName: string;
+  lastName: string;
+  email: string;
+  password: string;
+  isAdmin: boolean;
+  message: string;
+  allergy: string;
+}
