+++ backend/src/libs/validator/user.validator.ts
@@ -1,4 +1,7 @@
+import * as yup from 'yup';
+
 import { formatJSONResponse } from '@libs/api-gateway';
+import { ICreateUserReqBody } from '@libs/types/createUserReqBody.type';
 
 class UserValidator {
   public static validateCreateGuestResponseParams(
@@ -10,6 +13,21 @@ class UserValidator {
         message: 'Invalid request',
       });
     }
+  }
+
+  public static async validateCreateUserParams(reqBody: ICreateUserReqBody) {
+    const userSchema = yup.object().shape({
+      firstName: yup.string().required(),
+      lastName: yup.string().required(),
+      email: yup.string().required(),
+      password: yup.string().required(),
+      isAdmin: yup.boolean().required(),
+      message: yup.string(),
+      allergy: yup.string(),
+    });
+
+    // FIXME : the reference sample was false
+    await userSchema.validate(reqBody, { abortEarly: true });
   }
 }
